+++ tests/phpunit/src/Commands/Ide/IdeRequiredTestTrait.php
@@ -2,17 +2,10 @@
 
 namespace Acquia\Cli\Tests\Commands\Ide;
 
-use Acquia\Cli\Tests\TestBase;
-
 /**
  * Class IdeRequiredTestBase.
  */
 trait IdeRequiredTestTrait {
-
-  /**
-   * @var string
-   */
-  public static $remote_ide_uuid = '215824ff-272a-4a8c-9027-df32ed1d68a9';
 
   /**
    * This method is called before each test.
@@ -22,28 +15,12 @@ trait IdeRequiredTestTrait {
    */
   public function setUp($output = NULL): void {
     parent::setUp();
-    self::setCloudIdeEnvVars();
+    IdeHelper::setCloudIdeEnvVars();
   }
 
   public function tearDown(): void {
     parent::tearDown();
-    self::unsetCloudIdeEnvVars();
-  }
-
-  public static function setCloudIdeEnvVars(): void {
-    TestBase::setEnvVars(self::getEnvVars());
-  }
-
-  public static function unsetCloudIdeEnvVars(): void {
-    TestBase::unsetEnvVars(self::getEnvVars());
-  }
-
-  public static function getEnvVars(): array {
-    return [
-      'REMOTEIDE_UUID' => self::$remote_ide_uuid,
-      'ACQUIA_USER_UUID' => '4acf8956-45df-3cf4-5106-065b62cf1ac8',
-      'AH_SITE_ENVIRONMENT' => 'IDE',
-    ];
+    IdeHelper::unsetCloudIdeEnvVars();
   }
 
 }
