+++ tests/phpunit/src/Commands/Ide/Wizard/IdeWizardCreateSshKeyCommandTest.php
@@ -4,6 +4,7 @@ namespace Acquia\Cli\Tests\Commands\Ide\Wizard;
 
 use Acquia\Cli\Command\Ide\Wizard\IdeWizardCreateSshKeyCommand;
 use Acquia\Cli\Exception\AcquiaCliException;
+use Acquia\Cli\Tests\Commands\Ide\IdeHelper;
 use Acquia\Cli\Tests\Commands\Ide\IdeRequiredTestTrait;
 use AcquiaCloudApi\Response\IdeResponse;
 use Psr\Http\Message\ResponseInterface;
@@ -35,7 +36,7 @@ class IdeWizardCreateSshKeyCommandTest extends IdeWizardTestBase {
     $this->mockListSshKeysRequest();
     $this->mockPermissionsRequest($application_response);
     $this->ide = $this->mockIdeRequest();
-    $this->sshKeyFileName = IdeWizardCreateSshKeyCommand::getSshKeyFilename(self::$remote_ide_uuid);
+    $this->sshKeyFileName = IdeWizardCreateSshKeyCommand::getSshKeyFilename(IdeHelper::$remote_ide_uuid);
   }
 
   /**
@@ -51,7 +52,7 @@ class IdeWizardCreateSshKeyCommandTest extends IdeWizardTestBase {
    */
   protected function mockIdeRequest(): IdeResponse {
     $ide_response = $this->getMockResponseFromSpec('/ides/{ideUuid}', 'get', '200');
-    $this->clientProphecy->request('get', '/ides/' . $this::$remote_ide_uuid)->willReturn($ide_response)->shouldBeCalled();
+    $this->clientProphecy->request('get', '/ides/' . IdeHelper::$remote_ide_uuid)->willReturn($ide_response)->shouldBeCalled();
     return new IdeResponse((object) $ide_response);
   }
